Added P1's radio.mdl
With a few changes to work in Portal 2. It might reflect way too much light when it's directly facing the player. I'm making changes to the clipboard, mug and table.
Soon I can make the radio for all styles (except the Years Styles, since it will take more time and new models)